NewOrder separado.
Las clases newOrder fueron separadas, y se hicieron atributos estáticos para order y la lista details.
