+++ src/main/java/com/itec1/e_commerce/services/OrderServiceImpl.java
@@ -50,14 +50,9 @@ public class OrderServiceImpl {
         this.invoiceJpaController = new InvoiceJpaController(Connection.getEmf());
     }
 
-    public Order createOrder(Order entity) {
-        try {
-            orderJpaController.create(entity);
-            return orderJpaController.findOrder(entity.getId());
-        } catch (Exception e) {
-            System.err.println("Error while creating the order: " + e.getMessage());
-            throw new RuntimeException("Failed to create order.", e);
-        }
+    public Order createOrder(Order entity) throws Exception {
+        orderJpaController.create(entity);
+        return orderJpaController.findOrder(entity.getId());
     }
 
     public Order findById(Long id) {
@@ -182,7 +177,7 @@ public class OrderServiceImpl {
         detailOrderJpaController.edit(detail);
     }
 
-    public void qualifyCarrier(Invoice invoice, int star) throws Exception{
+    public void qualifyCarrier(Invoice invoice, int star) throws Exception {
         invoice.setCarrierQualification(star);
         invoiceJpaController.edit(invoice);
     }
